Add files via upload
Recorded on 15 June 2025 by Randall VK6WR:

There is a mystery signal on 40m that you can try your new Universal Radio Hacker skills on. It appears to be a FSK signal separated by 7kHz with the two signals at 7.0615 and 7.0685 MHz. Each of them on their own sounds a bit like a morse signal, but my CW decoder decodes junk. But if you can see it on a spectrum scope, it is clearly FSK because either one of them is on at any time.